Alterando tipo de dado de integer para BigDecimal para receber valores decimais

+++ Asaas/src/main/groovy/asaas/adapter/PaymentAdapter.groovy
@@ -10,7 +10,7 @@ import grails.compiler.GrailsCompileStatic
 @GrailsCompileStatic
 class PaymentAdapter {
 
-  Integer value
+  BigDecimal value
 
   Date dueDate
